+++ src/contents/settings/update.tsx
@@ -1,21 +1,18 @@
-import LogText from "@/components/log";
 import { Button } from "@/components/ui/button";
 import { INITIAL_UPDATE_STATE } from "@/lib/constants/states";
 import { IUpdatePageState } from "@/lib/types/states";
 import { cn } from "@/lib/utils";
 import { listen, UnlistenFn } from "@tauri-apps/api/event";
-import { AlertCircle, CheckCircle, RotateCw } from "lucide-react";
+import { AlertCircle, BugOff, CheckCircle, RotateCw } from "lucide-react";
 import { useEffect, useState } from "react";
 import { formatDistanceToNow } from "date-fns";
 import { Spinner } from "@/components/ui/spinner";
 import { invoke } from "@tauri-apps/api/core";
 import { getExitText, parseClamVersion } from "@/lib/helpers";
 import { toast } from "sonner";
-import { ScrollArea } from "@/components/ui/scroll-area";
-import { useSettings } from "@/context/settings";
+import SettingsItem from "@/components/settings-item";
 
-export default function UpdateContent(){
-     const {settings} = useSettings()
+export default function UpdateSettings(){
      const [updateState, setUpdateState] = useState<IUpdatePageState>(INITIAL_UPDATE_STATE);
      const [clamAvVersion, setClamavVersion] = useState<string>(()=>localStorage.getItem("clamav-version") || "");
      const setState = (overrides: Partial<IUpdatePageState>) => setUpdateState(prev=>({ ...prev, ...overrides }))
@@ -53,25 +50,11 @@ export default function UpdateContent(){
      useEffect(()=>{
           const unsubs: Promise<UnlistenFn>[] = [
                listen("freshclam:start",() => setState({
-                    log: [],
                     isUpdating: true
                })),
-               listen<string>("freshclam:output",e=>
+               listen("freshclam:error",()=>
                     setUpdateState(prev=>({
                          ...prev,
-                         log: [
-                              ...prev.log.slice(-settings.maxLogLines),
-                              e.payload
-                         ],
-                    }))
-               ),
-               listen<string>("freshclam:error",e=>
-                    setUpdateState(prev=>({
-                         ...prev,
-                         log: [
-                              ...prev.log.slice(-settings.maxLogLines),
-                              `ERROR: ${e.payload}`
-                         ],
                          isRequired: true,
                          isUpdating: false
                     }))
@@ -99,14 +82,16 @@ export default function UpdateContent(){
                Promise.all(unsubs).then(fns=>fns.forEach(fn=>fn()));
           }
      },[])
-     const {isRequired, exitMsg, log, lastUpdated, isUpdating} = updateState
+     const {isRequired, exitMsg, lastUpdated, isUpdating} = updateState
      const isInitializing = !lastUpdated && !isUpdating
      const Icon = (isUpdating || isInitializing) ? Spinner : !isRequired ? CheckCircle : AlertCircle
      return (
-          <>
-          <div className="space-y-4">
-               <h1 className="text-2xl md:text-3xl font-medium border-b pb-2 w-fit">Definition Updater</h1>
-               <div className="flex flex-col items-center gap-4">
+          <div className="px-1 py-2 space-y-3">
+               <SettingsItem
+                    Icon={BugOff}
+                    title="Definitions"
+                    className="flex flex-col items-center gap-4"
+               >
                     <div className="flex justify-center items-center gap-4">
                          <Icon className={cn("size-12",isRequired ? "text-destructive" : "text-emerald-600", (isUpdating || isInitializing) && "text-muted-foreground")}/>
                          <div className="text-center space-y-0.5">
@@ -133,14 +118,7 @@ export default function UpdateContent(){
                     {exitMsg!==null && (
                          <p className="text-sm text-muted-foreground">{getExitText(exitMsg,"update")}</p>
                     )}
-               </div>
+               </SettingsItem>
           </div>
-          <ScrollArea className="max-h-[800px]">
-               <div className="space-y-3 px-3 text-lg">
-                    <h2 className="text-2xl md:text-3xl font-medium border-b pb-2 w-fit">Log</h2>
-                    <LogText logs={log}/>
-               </div>
-          </ScrollArea>
-          </>
      )
 }