+++ src/lib/settings/tabs.ts
@@ -1,13 +1,12 @@
-import { ScanSettingsLoader, GeneralSettingsLoader, AdvancedSettingsLoader, ExclusionsSettingsLoader, SchedulerSettingsLoader } from "@/loaders/settings";
-import { Braces, ClipboardClock, Cog, Search, ShieldOff } from "lucide-react";
+import { ScanSettingsLoader, GeneralSettingsLoader, AdvancedSettingsLoader, UpdateSettingsLoader } from "@/loaders/settings";
+import { Braces, Cog, RotateCcw, Search } from "lucide-react";
 import { lazy } from "react";
 import { ISettingsTab } from "../types";
 
 const GeneralSettings = lazy(()=>import("@/contents/settings/general"));
 const AdvancedSettings = lazy(()=>import("@/contents/settings/advanced"));
 const ScanSettings = lazy(()=>import("@/contents/settings/scan"));
-const ExclusionsSettings = lazy(()=>import("@/contents/settings/exclusions"))
-const SchedulerSettings = lazy(()=>import("@/contents/settings/scheduler"))
+const UpdateSettings = lazy(()=>import("@/contents/settings/update"));
 
 export const SETTINGS_TABS: ISettingsTab[] = [
      {
@@ -25,24 +24,17 @@ export const SETTINGS_TABS: ISettingsTab[] = [
           LazyComponent: ScanSettings
      },
      {
-          page: "scheduler",
-          name: "Scheduler",
-          Icon: ClipboardClock,
-          Loader: SchedulerSettingsLoader,
-          LazyComponent: SchedulerSettings
-     },
-     {
-          page: "exclusions",
-          name: "Exclusions",
-          Icon: ShieldOff,
-          Loader: ExclusionsSettingsLoader,
-          LazyComponent: ExclusionsSettings
-     },
-     {
           page: "advanced",
           name: "Advanced",
           Icon: Braces,
           Loader: AdvancedSettingsLoader,
           LazyComponent: AdvancedSettings
      },
+     {
+          page: "update",
+          name: "Update",
+          Icon: RotateCcw,
+          Loader: UpdateSettingsLoader,
+          LazyComponent: UpdateSettings
+     }
 ]