+++ src/loaders/scan/scan-finish/with-threats.tsx
@@ -1,6 +1,6 @@
 import { Skeleton } from "@/components/ui/skeleton";
 import { TableLoaderProps } from "@/lib/types";
-import TableLoader from "@/loaders/components/table";
+import TableLoader from "@/components/loaders/table";
 
 export default function ScanFinishedTableLoader({rows}: TableLoaderProps){
      return (