+++ src/loaders/settings/exclusions.tsx
@@ -1,9 +1,0 @@
-import ExclusionsLoader from "../components/exclusions";
-
-export default function ExclusionsSettingsLoader(){
-     return (
-          <div className="px-1 py-2 space-y-3 w-full">
-               <ExclusionsLoader/>
-          </div>
-     )
-}