+++ src/loaders/settings/general.tsx
@@ -43,12 +43,31 @@ export default function GeneralSettingsLoader(){
                               </div>
                               <div className="w-[105px] h-9 bg-accent rounded-md"/>
                          </div>
-                         <div className="flex flex-row items-center justify-between w-full">
+                         {/* <div className="flex flex-row items-center justify-between w-full">
                               <div className="space-y-1 w-full">
                                    <div className="h-3.5 bg-accent rounded-md w-2/5"/>
                                    <div className="h-3.5 bg-accent rounded-md w-1/4"/>
                               </div>
                               <div className="w-32 h-9 bg-accent rounded-md"/>
+                         </div> */}
+                    </div>
+               </div>
+               <div className="border border-accent animate-pulse flex flex-col gap-5 rounded-xl p-6">
+                    <div className="h-4 bg-accent rounded-md w-1/3"/>
+                    <div className="space-y-4 w-full">
+                         <div className="flex flex-row items-center justify-between w-full">
+                              <div className="space-y-1 w-full">
+                                   <div className="h-3.5 bg-accent rounded-md w-2/5"/>
+                                   <div className="h-3.5 bg-accent rounded-md w-1/4"/>
+                              </div>
+                              <div className="w-8 h-[18px] bg-accent rounded-md"/>
+                         </div>
+                         <div className="flex flex-row items-center justify-between w-full">
+                              <div className="space-y-1 w-full">
+                                   <div className="h-3.5 bg-accent rounded-md w-2/5"/>
+                                   <div className="h-3.5 bg-accent rounded-md w-1/4"/>
+                              </div>
+                              <div className="w-8 h-[18px] bg-accent rounded-md"/>
                          </div>
                     </div>
                </div>