+++ src/loaders/settings/scan.tsx
@@ -1,5 +1,6 @@
 import { SCAN_SETTINGS_GROUPED } from "@/lib/settings/custom-scan-options";
 import { ScanOptionGroup } from "@/lib/types/settings";
+import ExclusionsLoader from "../../components/loaders/exclusions";
 
 export default function ScanSettingsLoader(){
      return (
@@ -8,20 +9,6 @@ export default function ScanSettingsLoader(){
                     <div className="h-4 bg-accent rounded-md w-1/3"/>
                     <div className="h-3.5 bg-accent rounded-md w-1/2"/>
                     <div className="space-y-4">
-                         <div className="flex flex-row items-center justify-between w-full">
-                              <div className="space-y-1 w-full">
-                                   <div className="h-3.5 bg-accent rounded-md w-2/5"/>
-                                   <div className="h-3.5 bg-accent rounded-md w-1/4"/>
-                              </div>
-                              <div className="w-8 h-[18px] bg-accent rounded-md"/>
-                         </div>
-                         <div className="flex flex-row items-center justify-between w-full">
-                              <div className="space-y-1 w-full">
-                                   <div className="h-3.5 bg-accent rounded-md w-2/5"/>
-                                   <div className="h-3.5 bg-accent rounded-md w-1/4"/>
-                              </div>
-                              <div className="w-8 h-[18px] bg-accent rounded-md"/>
-                         </div>
                          <div className="flex flex-row items-center justify-between w-full">
                               <div className="space-y-1 w-full">
                                    <div className="h-3.5 bg-accent rounded-md w-2/5"/>
@@ -54,6 +41,7 @@ export default function ScanSettingsLoader(){
                          </div>
                     </div>
                ))}
+               <ExclusionsLoader/>
           </div>
      )
 }